backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
requests
pymong[srv]
dnspython
